+++ template/AdminSAEStatistiques.php
@@ -15,6 +15,9 @@
         <td>
             &nbsp;
         </td>
+        <td>
+            &nbsp;
+        </td>
     </tr>
 
     <tr class="<?php echo $actes_nb_en_attente_sae_4h?"danger":"success" ?>">
@@ -24,6 +27,9 @@
             <a href="/modules/actes/index.php?status=<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
             </a>
+        </td>
+        <td>
+            &nbsp;
         </td>
     </tr>
 
@@ -37,6 +43,9 @@
                 Liste
             </a>
         </td>
+        <td>
+            &nbsp;
+        </td>
     </tr>
 
     <tr class="<?php echo $actes_erreur_lors_de_lenvoi_sae?"danger":"success" ?>">
@@ -47,6 +56,15 @@
                 Liste
             </a>
         </td>
+        <td>
+            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions')">
+                <input type="hidden" name="authority_id" value="<?php echo $authority_id ?>" />
+                <input type="hidden" name="status_id_from" value="<?php echo ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ENVOI_SAE ?>" />
+                <input type="hidden" name="status_id_to" value="<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>" />
+                <input type="submit" class="btn btn-warning" value="Relancer" />
+            </form>
+        </td>
+
     </tr>
     <tr class="<?php echo $actes_erreur_lors_de_larchivage?"danger":"success" ?>">
         <td>Actes erreur lors de l'archivage</td>
@@ -55,6 +73,16 @@
             <a href="/modules/actes/index.php?status=<?php echo ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ARCHIVAGE ?>&authority=<?php hecho($authority_id) ?>" class="icon">
                 Liste
             </a>
+        </td>
+        <td>
+            <?php if($actes_erreur_lors_de_larchivage > 0) : ?>
+            <form action="/modules/actes/actes_transac_change_status_sae_bulk.php" method="post" onsubmit="return confirm('Voulez-vous vraiment modifier ces transactions')">
+                <input type="hidden" name="authority_id" value="<?php echo $authority_id ?>" />
+                <input type="hidden" name="status_id_from" value="<?php echo ActesStatusSQL::STATUS_ERREUR_LORS_DE_L_ARCHIVAGE ?>" />
+                <input type="hidden" name="status_id_to" value="<?php echo ActesStatusSQL::STATUS_EN_ATTENTE_TRANMISSION_SAE ?>" />
+                <input type="submit" class="btn btn-warning" value="Relancer" />
+            </form>
+            <?php endif; ?>
         </td>
     </tr>